Made DietAssistantResults into a component and changed page.jsx to load the component.
